requirement.txt
fastapi
uvicorn
pydantic
python-dotenv
sqlalchemy
streamlit
